requirements/base.txt
# Core libraries shared across SegEdge pipelines
numpy>=1.24
tqdm>=4.66
PyYAML>=6.0
